requirements.txt
Flask
python-pptx
requests
openai
